pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
]
